Added certificate cache override
These changes create a certificate cache override on the Webscale TLS
config, formerly certmagic. This override can be set with an alias-keyed
certificate cache and, when set, is used instead of the default
certificate-name-keyed cache. The alias-keyed cache will use the
SNI from the client's hello message during the TLS handshake to look for
an exact match in the certificate cache. If one is found, its
certificate is returned. Otherwise, a default certificate is returned.