Added certificate cache override
These changes create an alias-keyed TLS certificate cache that lookups a
certificate to use during the TLS handshake based only on the server
name in the TLS client's hello message. These alias-certificate
associations are explicitly added by Webscale's 'strict_tls' Caddyfile
directive.

A short-circuit was added in the TLS certificate lookup that uses an
alias-keyed cache instead of certmagic's implementation when the
configuration's 'OverrideCache' field is set. This certificate cache
differs from certmagic's implementation as it does not use the names
within a certificate at all.